fix(ci): python-coverage-shards - split oversized executor and recovery test files
